requirements.txt
absl-py==2.3.1
accelerate==1.8.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.41.1
blessed==1.21.0
Brotli==1.1.0
bupy==0.1.2
bypy==1.8.9
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
clip==0.2.0
colorama==0.4.6
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.1
cycler==0.12.1
decorator==5.2.1
deepspeed==0.12.6
dill==0.4.0
dropblock==0.3.0
einops==0.4.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.100.1
ffmpy==0.6.0
filelock==3.14.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
gpustat==1.1.1
gradio==3.39.0
gradio_client==1.10.4
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
inflate64==1.0.3
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
linkify-it-py==2.0.3
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.10.3
matplotlib-inline==0.2.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmcv==2.2.0
mmdet==3.3.0
mmengine==0.10.7
mmpretrain==1.2.0
mmsegmentation==1.2.2
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.18
multivolumefile==0.2.3
narwhals==1.45.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.24.2
nvidia-ml-py==12.575.51
nvitop==1.5.2
omegaconf==2.3.0
openai==0.27.8
opencv-python==4.8.0.74
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.18
oss2==2.17.0
packaging==24.2
pandas==2.3.0
parso==0.8.5
peft==0.4.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==1.0.0
pybcj==1.0.6
pycocotools==2.0.6
pycparser==2.23
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pynvml==12.0.0
pyparsing==3.2.3
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2023.4
pyxdg==0.28
PyYAML==6.0.2
pyzstd==0.18.0
ray==2.6.1
referencing==0.36.2
regex==2024.11.6
requests==2.28.2
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.26.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.14
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
shortuuid==1.0.11
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
texttable==1.7.0
tifffile==2025.5.10
timm==1.0.20
tokenizers==0.13.3
tomli==2.2.1
torch==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.65.2
traitlets==5.14.3
transformers==4.31.0
triton==2.1.0
typer==0.9.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==1.26.20
uvicorn==0.23.2
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
